+++ docs/CHANGES.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## 2.0.6 (2017-02-01)
+## 2.0.6 (Unreleased)
 - #486 Honor existing mam user configuration [throwaway42]
+- Escape user-generated input to prevent JS-injection attacks. (Thanks to SamWhited) [jcbrand]
 
 ## 2.0.5 (2017-02-01)
 - #743, #751, #753 Update to Strophe 1.2.12. SASL-EXTERNAL now has reduced priority, so it won't